change workerID to participantID
This change looks innocous, but could cause problems down the line. My
fear is that the URL parameter is used somewhere unexpected.

As far as I can tell, the URL parameter is only used in task.ts, where
by great coincidence, it looks for "participantID" as well, so I don't
even need to change anything in task.ts

The userID is stored, but it seems like it is stored in a session that
is unrelated to the URL parameter. Copying the cartoscope link, closing
the tab, clearing your history, and re-opening the tab does break the
link, but it seems like that bug existed before this commit anyways.